requirement.txt
Flask
numpy
pandas
python-dotenv
mysql-connector-python
pymysql
scikit-learn